requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0
jinja2>=3.1.0


# Messaging (Обмін повідомленнями)
pika==1.3.2

# Environment configuration (Налаштування оточення)
python-dotenv>=1.0.0

# Testing (Тестування)
pytest>=7.0.0
httpx>=0.25.0

# Persistence (Збереження даних)
sqlmodel>=0.0.14
alembic>=1.13.0


# PostgreSQL driver (Драйвер PostgreSQL)
psycopg[binary]==3.2.3

# Graph visualization (Візуалізація графів)
graphviz>=0.20.3

# FastAPI forms (Підтримка форм)
python-multipart>=0.0.9

